[svn-r6109] *** empty log message ***

+++ test/tattr.c
@@ -186,9 +186,9 @@ test_attr_basic_write(void)
     CHECK(attr_name_size, FAIL, "H5Aget_name");
     
     if(attr_name_size>0)
-        attr_name = (char*)HDcalloc(attr_name_size+1, sizeof(char));
+        attr_name = (char*)HDcalloc((size_t)(attr_name_size+1), sizeof(char));
     
-    ret=H5Aget_name(attr, attr_name_size+1, attr_name);
+    ret=H5Aget_name(attr, (size_t)(attr_name_size+1), attr_name);
     CHECK(ret, FAIL, "H5Aget_name");
     ret=HDstrcmp(attr_name, ATTR_TMP_NAME);
     VERIFY(ret, 0, "HDstrcmp");
@@ -221,9 +221,9 @@ test_attr_basic_write(void)
     CHECK(attr_name_size, FAIL, "H5Aget_name");
     
     if(attr_name_size>0)
-        attr_name = (char*)HDcalloc(attr_name_size+1, sizeof(char));
+        attr_name = (char*)HDcalloc((size_t)(attr_name_size+1), sizeof(char));
     
-    ret=H5Aget_name(attr2, attr_name_size+1, attr_name);
+    ret=H5Aget_name(attr2, (size_t)(attr_name_size+1), attr_name);
     CHECK(ret, FAIL, "H5Aget_name");
     ret=HDstrcmp(attr_name, ATTR1A_NAME);
     VERIFY(ret, 0, "HDstrcmp");
